+++ packages/web/hooks/index.ts
@@ -4,3 +4,4 @@ export * from "./use-boolean-with-window-event";
 export * from "./use-ibc-transfer";
 export * from "./use-amount-config";
 export * from "./use-fake-fee-config";
+export * from "./window";
