+++ packages/web/hooks/window/index.ts
@@ -1,0 +1,2 @@
+export * from "./use-window-scroll";
+export * from "./use-window-size";
